The Big Fucking Fix - Part 4

+++ Assets/Scripts/PlayerController.cs
@@ -18,8 +18,8 @@ public class PlayerController : MonoBehaviour
     public float upgradeMaxHorizontalSpeed = 3.0f;
 
     [Header("==== SCORE VALUES ====")]
-    public float currentOreValue;
-    public float currentMoney;
+    public long currentOreValue;
+    public long currentMoney;
 
     [Header("==== Vertical Data ====")]
     public float currentAngle = 0f;
